feat/complete-nnx
